Rename async module and fix workflow
